create new assessment

+++ app/src/main/java/com/revature/caliberdroid/data/parser/JSONParser.kt
@@ -82,7 +82,7 @@ object JSONParser {
                     getString("assessmentType"),
                     getInt("weekNumber"),
                     getLong("batchId"),
-                    getInt("assessmentCategory"))
+                    getLong("assessmentCategory"))
             }
             assessmentList.add(assessment)
         }
